[WIP] Getting local ApiAuthenticationData on startup.
This causes an error with ERPLibre. Current theories as to why this
happens include encoding errors or the API token not being passed. This
error was previously encountered when sending a request after
subscribing to the observable that requested an API token from ERPLibre
instead of sending it in the next() method of the observer.