[FIX] Reflet: Element height calculations
Since we need the chat box to have a scrollbar while also having a
dynamic height, we need to give it a height or max-height CSS style at
least, to prevent the content from making the chat box larger than the
desired height. This commit fixes previous calculations and prevents the
reflet component from being larger than its desired maximum height.
Without this fix, you would be able to scroll down in the reflet
component, with a scrollbar that would allow the user to scroll only a
couple pixels at most in certain situations. It's only an annoyance, but
I'd rather fix it now than having to deal with it later.