[UPD] RefletChatComponent: Improved chat box scroll speed
After making some changes to the way of fetching personalChatInformation
in our reflet components, the MutationObserver for messageList stopped
working. The added nodes were not picked up by the observer anymore.
That means the observer starts observing after the data is gathered by
fetchPersonalChatInformation and put in the messageList HTMLElement. The
fix for this was found by accident, but it works better than the
previous solution, and is on par with the original solution in terms of
visual performance. The mutation cannot detect the nodes added to
messageList, but it can detect the style attribute being added to it
when we calculate the max-height of messageList. This update eliminates
the visible latency between the chat messages appearing on screen and
the instant scroll to the bottom of messageList.